requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0

# Anthropic SDK (direct usage)
anthropic>=0.40.0

# Vector Store
chromadb>=0.5.0

# Data validation
pydantic>=2.0.0

# Web UI
streamlit>=1.40.0

# Environment management
python-dotenv>=1.0.0

# Image Generation
google-genai>=1.0.0

# Utilities
uuid6>=2024.1.12
